refactored db to have a clear API :3
