+++ handlers/web/CreateCredentials.go
@@ -24,7 +24,7 @@ func (h *Handler) CreateCredentials(c *gin.Context) {
 		return
 	}
 
-	accessKey, secretKey, err := h.Store.CreateCredentials(request.Name, request.Description, request.ExpiresAt, false, id.(int64))
+	accessKey, secretKey, err := h.Store.PutCredentials(request.Name, request.Description, request.ExpiresAt, false, id.(int64))
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
