--- storage/db/DeleteCredentials.go
@@ -1,6 +1,0 @@
-package db
-
-func (s *Store) DeleteCredentials(accessKey string) error {
-	_, err := s.db.Exec(`DELETE FROM credentials WHERE access_key = ?`, accessKey)
-	return err
-}
