+++ storage/db/db.go
@@ -2,140 +2,172 @@ package db
 
 import (
 	"database/sql"
-	"fmt"
+	"mini-alt/models"
+	"mini-alt/types"
 )
 
 type Store struct {
 	db *sql.DB
 }
 
-func NewSQLiteStore(dbPath string) (*Store, error) {
-	db, err := sql.Open("sqlite3", dbPath)
-	if err != nil {
-		return nil, err
-	}
+// ----------------- Management --------------
 
-	_, err = db.Exec(`PRAGMA foreign_keys = ON`)
-	if err != nil {
-		return nil, err
-	}
-
-	store := &Store{db: db}
-	if err := store.initSchema(); err != nil {
-		return nil, err
-	}
-	if err := store.SeedInitialData(); err != nil {
-		return nil, err
-	}
-
-	return store, nil
+func NewStore(path string) (*Store, error) {
+	return newSQLiteStore(path)
 }
 
-func (s *Store) initSchema() error {
-	schema := `
-	CREATE TABLE IF NOT EXISTS buckets (
-		id INTEGER PRIMARY KEY AUTOINCREMENT,
-		name TEXT UNIQUE NOT NULL,
-		owner INTEGER,
-		created_at DATETIME NOT NULL
-	);
+// ----------------- Buckets -----------------
 
-	CREATE TABLE IF NOT EXISTS objects (
-		id INTEGER PRIMARY KEY AUTOINCREMENT,
-		bucket_name TEXT NOT NULL,
-		key TEXT NOT NULL,
-		size INTEGER,
-		last_modified DATETIME,
-		FOREIGN KEY(bucket_name) REFERENCES buckets(name) ON DELETE CASCADE,
-		UNIQUE(bucket_name, key)
-	);
-
-	CREATE INDEX IF NOT EXISTS idx_objects_bucket_name ON objects(bucket_name);
-
-	CREATE TABLE IF NOT EXISTS object_metadata (
-	    id INTEGER PRIMARY KEY AUTOINCREMENT,
-	    object_id INTEGER NOT NULL UNIQUE,
-	    acl TEXT,
-	    cache_control TEXT,
-	    content_disposition TEXT,
-	    content_encoding TEXT,
-	    content_language TEXT,
-	    content_length INTEGER,
-	    content_md5 TEXT,
-	    content_type TEXT,
-	    expires DATETIME,
-	    FOREIGN KEY(object_id) REFERENCES objects(id) ON DELETE CASCADE
-	);
-
-	CREATE TABLE IF NOT EXISTS  credentials (
-		id INTEGER PRIMARY KEY AUTOINCREMENT,
-		access_key TEXT UNIQUE NOT NULL,
-		secret_key_encrypted TEXT NOT NULL,
-		user BOOLEAN NOT NULL,
-		expires_at DATE DEFAULT NULL,
-		status BOOLEAN NOT NULL DEFAULT TRUE,
-		name TEXT,
-		description TEXT,
-		owner INTEGER NOT NULL,
-		created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-	);
-
-	CREATE TABLE IF NOT EXISTS users (
-	    id INTEGER PRIMARY KEY AUTOINCREMENT,
-	    username TEXT UNIQUE NOT NULL UNIQUE,
-	    password TEXT NOT NULL,
-	    token TEXT NOT NULL,
-		access_key TEXT NOT NULL,
-		admin BOOLEAN NOT NULL DEFAULT FALSE,
-	    expires_at DATE DEFAULT NULL,
-	    created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-	);
-	
-	CREATE TABLE IF NOT EXISTS config (
-	    id INTEGER PRIMARY KEY AUTOINCREMENT,
-	    key TEXT UNIQUE NOT NULL,
-	    value TEXT NOT NULL,
-	    created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-	);
-
-	CREATE TABLE IF NOT EXISTS events (
-	    id INTEGER PRIMARY KEY AUTOINCREMENT,
-	    name TEXT UNIQUE NOT NULL,
-	    description TEXT,
-	    bucket_id INTEGER NOT NULL,
-	    endpoint TEXT NOT NULL,
-	    token TEXT,
-	    created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-	);
-`
-	_, err := s.db.Exec(schema)
-	return err
+func (s *Store) GetBucket(bucket string) (models.Bucket, error) {
+	return s.getBucket(bucket)
 }
 
-func (s *Store) SeedInitialData() error {
-	var count int
-	err := s.db.QueryRow(`SELECT COUNT(*) FROM users`).Scan(&count)
-	if err != nil {
-		return fmt.Errorf("failed to count users: %w", err)
-	}
+func (s *Store) GetBucketById(id int64) (models.Bucket, error) {
+	return s.getBucketById(id)
+}
 
-	if count > 0 {
-		return nil
-	}
-	accessKey, _, err := s.CreateCredentials("", "", "", true, 1)
-	if err != nil {
-		return err
-	}
+func (s *Store) PutBucket(bucket string, owner int64) error {
+	return s.putBucket(bucket, owner)
+}
 
-	id, err := s.RegisterUser("admin", "admin", accessKey, "", true)
-	if err != nil {
-		return fmt.Errorf("failed to insert default user: %w", err)
-	}
+func (s *Store) ListBuckets(owner int64) ([]models.Bucket, error) {
+	return s.listBuckets(owner)
+}
 
-	err = s.AddCredentialsOwner(accessKey, id)
-	if err != nil {
-		return fmt.Errorf("failed to insert default user: %w", err)
-	}
+func (s *Store) DeleteBucket(bucket string) error {
+	return s.deleteBucket(bucket)
+}
 
-	return nil
+func (s *Store) AddBucketOwner(bucketName string, bucketOwner int64) error {
+	return s.addBucketOwner(bucketName, bucketOwner)
+}
+
+// ----------------- Users -------------------
+
+func (s *Store) GetUser(username string) (models.User, error) {
+	return s.getUser(username)
+}
+
+func (s *Store) GetUserById(id int64) (models.User, error) {
+	return s.getUserById(id)
+}
+
+func (s *Store) GetUserByAccessKey(accessKey string) (models.User, error) {
+	return s.getUserByAccessKey(accessKey)
+}
+
+func (s *Store) ListUsers() ([]models.User, error) {
+	return s.listUsers()
+}
+
+func (s *Store) DeleteUser(id int64) error {
+	return s.deleteUser(id)
+}
+
+func (s *Store) DeleteExpiredUsers() {
+	s.deleteExpiredUsers()
+}
+
+func (s *Store) AuthenticateAdmin(id int64, token string) error {
+	return s.authenticateAdmin(id, token)
+}
+
+func (s *Store) AuthenticateUser(id int64, token string) error {
+	return s.authenticateUser(id, token)
+}
+
+func (s *Store) LoginUser(username, password string) (LoginResponse, error) {
+	return s.loginUser(username, password)
+}
+
+func (s *Store) RegisterUser(username, password, accessKey, expiresAt string, admin bool) (int64, error) {
+	return s.registerUser(username, password, accessKey, expiresAt, admin)
+}
+
+// ----------------- Objects -----------------
+
+func (s *Store) GetObject(bucket, key string) (models.Object, error) {
+	return s.getObject(bucket, key)
+}
+
+func (s *Store) ListObjects(bucket string) ([]models.Object, error) {
+	return s.listObjects(bucket)
+}
+
+func (s *Store) PutObject(bucket, object string, size int64) (models.Object, error) {
+	return s.putObject(bucket, object, size)
+}
+
+func (s *Store) DeleteObject(bucket, key string) error {
+	return s.deleteObject(bucket, key)
+}
+
+// ----------------- Metadata ----------------
+
+func (s *Store) GetMetadata(objectId int64) (models.ObjectMetadata, error) {
+	return s.getMetadata(objectId)
+}
+
+func (s *Store) PutMetadata(objectId int64, metadata types.Metadata) error {
+	return s.putMetadata(objectId, metadata)
+}
+
+func (s *Store) CopyMetadata(oldObjectId, newObjectId int64) error {
+	return s.copyMetadata(oldObjectId, newObjectId)
+}
+
+// ----------------- Credentials -------------
+
+func (s *Store) GetSecretKey(accessKey string) (string, error) {
+	return s.getSecretKey(accessKey)
+}
+
+func (s *Store) ListCredentials(owner int64) ([]models.Credentials, error) {
+	return s.listCredentials(owner)
+}
+
+func (s *Store) PutCredentials(name, description, expiresAt string, user bool, owner int64) (accessKey, secretKey string, err error) {
+	return s.putCredentials(name, description, expiresAt, user, owner)
+}
+
+func (s *Store) DeleteCredentials(accessKey string) error {
+	return s.deleteCredentials(accessKey)
+}
+
+func (s *Store) DeleteExpiredCredentials() {
+	s.deleteExpiredCredentials()
+}
+
+func (s *Store) AddCredentialsOwner(accessKey string, owner int64) error {
+	return s.addCredentialsOwner(accessKey, owner)
+}
+
+func (s *Store) EditCredentials(accessKey string, name, description, expiresAt string, status bool) error {
+	return s.editCredentials(accessKey, name, description, expiresAt, status)
+}
+
+// ----------------- Events ------------------
+
+func (s *Store) ListEvents() ([]models.Event, error) {
+	return s.listEvents()
+}
+
+func (s *Store) CreateEvent(name, description, endpoint, token string, bucket int64) error {
+	return s.createEvent(name, description, endpoint, token, bucket)
+}
+
+// ----------------- Config ------------------
+
+func GetConfig[T any](s *Store, key string) (T, error) {
+	return getConfig[T](s, key)
+}
+
+func PutConfig(s *Store, key string, value interface{}) error {
+	return putConfig(s, key, value)
+}
+
+// ----------------- Server Information ------
+
+func (s *Store) GetServerInformation() (*ServerInformation, error) {
+	return s.getServerInformation()
 }
