+++ storage/db/event.go
@@ -1,10 +1,8 @@
 package db
 
-import (
-	"mini-alt/models"
-)
+import "mini-alt/models"
 
-func (s *Store) ListEvents() ([]models.Event, error) {
+func (s *Store) listEvents() ([]models.Event, error) {
 	rows, err := s.db.Query(`SElECT * FROM events`)
 	if err != nil {
 		return nil, err
@@ -21,4 +19,13 @@ func (s *Store) ListEvents() ([]models.Event, error) {
 	}
 
 	return events, nil
+}
+
+func (s *Store) createEvent(name, description, endpoint, token string, bucket int64) error {
+	_, err := s.db.Exec(`INSERT INTO events (name, description, endpoint, token, bucket_id) VALUES (?, ?, ?, ?, ?)`, name, description, endpoint, token, bucket)
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
